requirements.txt
flask
Flask-CORS==4.0.0
streamlit
extra_streamlit_components
pyjwt
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
boto3
botocore
aws-opentelemetry-distro
ddgs
aws-opentelemetry-distro~=0.10.1
pyyaml